requirements.txt
Django==3.2.5
djangorestframework==3.12.4
pandas==1.1.5

